% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condpath.R
\name{condtour}
\alias{condtour}
\alias{randomPath}
\alias{seqPath}
\alias{alongPath}
\alias{kmeansPath}
\alias{pamPath}
\alias{claraPath}
\alias{lofPath}
\alias{diffitsPath}
\alias{createPath}
\alias{centroidPath}
\alias{medoidPath}
\title{Tours of data space}
\usage{
randomPath(
  data,
  fits = NULL,
  length = 10,
  reorder = TRUE,
  conditionvars = NULL,
  ...
)

seqPath(
  data,
  fits = NULL,
  length = 10,
  reorder = FALSE,
  conditionvars = NULL,
  ...
)

alongPath(data, var, length = 10, current = NULL, ...)

kmeansPath(
  data,
  fits = NULL,
  length = 10,
  reorder = TRUE,
  conditionvars = NULL,
  ...
)

pamPath(
  data,
  fits = NULL,
  length = 10,
  reorder = TRUE,
  conditionvars = NULL,
  maxn = 4000,
  ...
)

claraPath(
  data,
  fits = NULL,
  length = 10,
  reorder = TRUE,
  conditionvars = NULL,
  ...
)

lofPath(
  data,
  fits,
  length = 10,
  reorder = TRUE,
  conditionvars = NULL,
  predictArgs = NULL,
  response = NULL,
  ...
)

diffitsPath(
  data,
  fits,
  length = 10,
  reorder = TRUE,
  conditionvars = NULL,
  predictArgs = NULL,
  ...
)

createPath(data, score, length = 10, reorder = TRUE, conditionvars = NULL)

centroidPath(data, cl, reorder = FALSE)

medoidPath(data, cl, reorder = FALSE)
}
\arguments{
\item{data}{A dataset}

\item{fits}{A model fit or list of fits}

\item{length}{The length of path returns}

\item{reorder}{If TRUE, points on the path are re-ordered so nearby points are close in the path}

\item{conditionvars}{A vector of variable names. Some tours will limit calculation to this subset of variables.}

\item{...}{other arguments, ignored}

\item{maxn}{If data has more than this number of rows, use maxn sample of rows for speed reasons (pamPath only)}

\item{predictArgs}{a list with one entry per fit, giving arguments for CVpredict}

\item{score}{A vector of length equal to the nrows of data.}
}
\value{
A dataframe, which is the path
}
\description{

}
\details{
kmeansPath works for both numeric and factors, which are converted to columns of indicators.
pamPath uses a subset of rows for large datasets, see maxn.
fits are used only in lofPath and diffitsPath.
Paths are reordered using dser from package DendSer.


}
\section{Functions}{
\itemize{
\item \code{randomPath}: Returns a random path

\item \code{seqPath}: Returns a  path of first length cases

\item \code{alongPath}: Returns a  path along var

\item \code{kmeansPath}: Returns a path using kmeans centroids

\item \code{pamPath}: Returns a path using pam medoids from package cluster

\item \code{claraPath}: Returns a path using clara medoids from package cluster

\item \code{lofPath}: Returns a path showing biggest absolute residuals from fits.

\item \code{diffitsPath}: Returns a path showing biggest difference in fits

\item \code{createPath}: Returns a path showing highest scores

\item \code{centroidPath}: Returns a path visiting cluster centroids

\item \code{medoidPath}: Returns a path visiting cluster medoids
}}

