% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUI.R
\name{createCVUI}
\alias{createCVUI}
\title{Constructs UI for Condvis}
\usage{
createCVUI(
  CVfit,
  data,
  response,
  sectionvars,
  preds = NULL,
  pointColor,
  threshold = 1,
  thresholdmax,
  tours,
  probs,
  view3d,
  showsim,
  cPlotPCP
)
}
\arguments{
\item{CVfit}{a list of fits}

\item{data}{a dataset}

\item{response}{name of response variable}

\item{sectionvars}{names of sectionvars}

\item{preds}{names of predictors}

\item{pointColor}{a color, or the name of variable to be used for coloring}

\item{threshold}{used for similarity weights, defaults to 1.}

\item{thresholdmax}{maximum value allowed of threshold.}

\item{tours}{A list of pre-calculated tours}

\item{probs}{Logical; if \code{TRUE}, shows predicted class probabilities instead of just predicted classes.}

\item{view3d}{Logical; if \code{TRUE}, includes option for a three-dimensional regression surface if possible.}

\item{showsim}{if TRUE, shows sim in conditionplots with points/lines. Defaults to TRUE with 150 or fewer cases.}

\item{cPlotPCP}{if TRUE, conditionplots are drawn as a single PCP (for more than two conditionvars)}
}
\value{
a dataframe of conditions
}
\description{
Constructs UI for Condvis
}
