% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condpath.R
\name{medoid}
\alias{medoid}
\title{Finds medoid of data}
\usage{
medoid(data, maxn = 4000)
}
\arguments{
\item{data}{}

\item{maxn}{If data has more than this number of rows, use maxn sample of rows for speed reasons}
}
\value{
A dataframe with one row, which is the medoid of the data, based on (standardised) daisy dist
}
\description{
Finds medoid of data
}
