% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plotTourDiagnostics}
\alias{plotTourDiagnostics}
\title{Plots diagnostics for the tour supplied}
\usage{
plotTourDiagnostics(path, data, pathlen = nrow(path), threshold = 1,
  which = 1:3, ...)
}
\arguments{
\item{path}{the tour}

\item{data}{the dataset}

\item{pathlen}{the pathlength}

\item{threshold}{used for similarityweight}

\item{which}{subset of 1:3}

\item{...}{other args for similarityweight}
}
\value{
Table of max sims attained.
}
\description{
Plots diagnostics for the tour supplied
}
\details{
The first plot shows approximately how much data are visible on each section,  the second shows what proportion of data are \emph{visited} by the tour, and the third a density estimate of max similarity values.
}
