% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sectionplots.R
\name{sectionPlot}
\alias{sectionPlot}
\title{Plots the main condvis display}
\usage{
sectionPlot(CVdata, CVfit, response, preds, sectionvar, conditionvals,
  pointColor = NULL, sim = NULL, threshold = 1, linecols = NULL,
  dataplot = NULL, gridsize = 50, probs = FALSE, view3d = FALSE,
  theta3d = 45, phi3d = 20, xlim = NULL, ylim = NULL,
  zlim = NULL, pointSize = 2, predictArgs = NULL, resetpar = TRUE,
  density = FALSE, showdata = density == FALSE, returnCoords = FALSE)
}
\arguments{
\item{CVdata}{the dataset used for the fit}

\item{CVfit}{a fit or list of fits}

\item{response}{name of response variable}

\item{preds}{names of predictors}

\item{sectionvar}{section variable}

\item{conditionvals}{conditioning values. A list or dataframe with one row}

\item{pointColor}{a color, or the name of variable to be used for coloring}

\item{sim}{vector of similarity weights}

\item{threshold}{used for similarity weights, defaults to 1.}

\item{linecols}{vector of line colours}

\item{dataplot}{if CVfit is NULL, the data are plotted using this function. Defaults to a parallel coordinate plot}

\item{gridsize}{used to construct grid of fitted values.}

\item{probs}{Logical; if \code{TRUE}, shows predicted class probabilities instead of just predicted classes. Only available with two numeric sectionvars and the model's predict method provides this.}

\item{view3d}{Logical; if \code{TRUE} plots a three-dimensional regression surface if possible.}

\item{theta3d, phi3d}{Angles defining the viewing direction. \code{theta3d} gives the azimuthal direction and \code{phi3d} the colatitude. See\code{\link[graphics]{persp}}.}

\item{xlim}{passed on to plot}

\item{ylim}{passed on to plot}

\item{zlim}{passed on to plot}

\item{pointSize}{used for points}

\item{predictArgs}{a list with one entry per fit, giving arguments for predict}

\item{resetpar}{When TRUE (the default) resets pars after drawing.}

\item{density}{default FALSE. Use TRUE if model is a density function.}

\item{showdata}{If FALSE, data on section not shown.}

\item{returnCoords}{If TRUE, returns coordinates for some plots}
}
\value{
plotted coordinates, for some plots
}
\description{
Plots the main condvis display
}
