% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{weightcolor}
\alias{weightcolor}
\title{Fade colours according to a weight vector}
\usage{
weightcolor(col, weights, nlevels = 5)
}
\arguments{
\item{col}{A vector of colour}

\item{weights}{A vector of weights, values between 0 and 1}

\item{nlevels}{The number of groups}
}
\value{
A vector of colours
}
\description{
The colours whose weights are less than 1 are diluted. Colours whose weight is zero are returned as white, 
other weights are grouped in \code{nlevels} groups and colours diluted proportionally.
}
