% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluspath.R
\name{medoid}
\alias{medoid}
\title{Finds medoid of data}
\usage{
medoid(data, maxn = 4000)
}
\arguments{
\item{data}{A dataframe}

\item{maxn}{For datasets with more than maxn rows, use maxn randomly selected rows.}
}
\value{
A dataframe with one row, which is the medoid of the data, based on (standardised) daisy dist
}
\description{
Finds medoid of data
}
\examples{
medoid(mtcars)

}
