% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condpath.R
\name{pathInterpolate}
\alias{pathInterpolate}
\alias{pathInterpolate.default}
\alias{pathInterpolate.factor}
\alias{pathInterpolate.data.frame}
\title{Interpolation}
\usage{
pathInterpolate(x, ninterp = 4)

\method{pathInterpolate}{default}(x, ninterp = 4L)

\method{pathInterpolate}{factor}(x, ninterp = 4L)

\method{pathInterpolate}{data.frame}(x, ninterp = 4L)
}
\arguments{
\item{x}{a numeric or factor vector or dataframe}

\item{ninterp}{number of interpolated steps}
}
\value{
interpolated version of x
}
\description{
Interpolation
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default interpolate method

\item \code{factor}: pathInterpolate method for factor

\item \code{data.frame}: pathInterpolate method for data.frame
}}

